added first cut of service container
