+++ go.mod
@@ -3,12 +3,14 @@ module github.com/kgoins/ldsview
 go 1.13
 
 require (
+	github.com/armon/consul-api v0.0.0-20180202201655-eb2c6b5be1b6 // indirect
 	github.com/audibleblink/msldapuac v0.2.0
 	github.com/fsnotify/fsnotify v1.4.9 // indirect
 	github.com/kgoins/entityfilter v1.0.0
 	github.com/kgoins/hashset v0.1.0
 	github.com/kgoins/ldapentity v0.1.0
 	github.com/kgoins/ldifparser v0.2.1
+	github.com/kgoins/snakecharmer v0.1.0
 	github.com/magiconair/properties v1.8.4 // indirect
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/mitchellh/mapstructure v1.3.3 // indirect
@@ -16,11 +18,13 @@ require (
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/sarulabs/di v2.0.0+incompatible
 	github.com/spf13/afero v1.4.0 // indirect
-	github.com/spf13/cobra v1.0.0
+	github.com/spf13/cobra v1.1.1
 	github.com/spf13/jwalterweatherman v1.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/spf13/viper v1.7.1
 	github.com/stretchr/testify v1.7.0
+	github.com/ugorji/go v1.1.4 // indirect
+	github.com/xordataexchange/crypt v0.0.3-0.20170626215501-b2862e3d0a77 // indirect
 	go.uber.org/atomic v1.8.0 // indirect
 	go.uber.org/multierr v1.7.0 // indirect
 	go.uber.org/zap v1.18.1
